# api-doc.yml
swagger: "2.0"
info:
  description: ""
  version: "0.1.0"
  title: "CMRS 接口描述文档"
  termsOfService: "http://117.78.6.206:9580"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "139.159.162.133:9580"
basePath: "/"
...
schemes:
  - http
security:
  - CMRSSecurity: []
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /create-restaurant:
    post:
      tags:
        - "管理员维护餐厅的基本信息"
      summary: "管理员创建餐厅接口"
      parameters:
        - in: body
          name: "管理员创建餐厅命令"
          required: true
          schema:
            $ref: "#/definitions/CreateRestaurantCommand"
      responses:
        200:
          description: "创建餐厅成功"
          schema:
            $ref: "#/definitions/AggregateCreatedDTO"
  /update-details-of-restaurant/:restId:
    post:
      tags:
        - "管理员维护餐厅的基本信息"
      summary: "管理员更新餐厅接口"
      parameters:
        - name: "restId"
          type: string
          in: path
          required: true
          description: "餐厅ID"
        - name: "管理员更新餐厅命令"
          in: body
          required: true
          schema:
            $ref: "#/definitions/UpdateRestaurantCommand"
      responses:
        204:
          description: "更新成功"
...
  /create-single-menu-item:
    post:
      tags:
        - "管理员维护餐厅的菜单信息"
      summary: "创建单品菜单项接口"
      parameters:
        - in: body
          name: "创建单品菜单命令"
          required: true
          schema:
            $ref: "#/definitions/CreateSingleMenuItemCommand"
      responses:
        200:
          description: "创建单品菜单项成功"
          schema:
            $ref: "#/definitions/AggregateCreatedDTO"
  /create-set-menu-item:
    post:
      tags:
        - "管理员维护餐厅的菜单信息"
      summary: "创建套餐菜单项接口"
      parameters:
        - name: "创建单品菜单命令"
          in: body
          required: true
          schema:
            $ref: "#/definitions/CreateSetMenuItemCommand"
      responses:
        200:
          description: "创建套餐菜单项成功"
          schema:
            $ref: "#/definitions/AggregateCreatedDTO"
  /update-restaurant/:restId/single-menu-item/:itemId:
    post:
      tags:
        - "管理员维护餐厅的菜单信息"
      summary: "更新单品菜单项接口"
      parameters:
        - name: "restId"
          in: path
          required: true
          type: string
          description: "菜单项所属菜单餐厅ID"
        - name: "itemId"
          in: path
          required: true
          type: string
          description: "菜单项ID"
        - name: "更新单品菜单项命令"
          in: body
          required: true
          schema:
            $ref: "#/definitions/UpdateSingleMenuItemCommand"
      responses:
        204:
          description: "更新单品菜单项成功"
  /update-restaurant/:restId/set-menu-item/:itemId:
    post:
      tags:
        - "管理员维护餐厅的菜单信息"
      summary: "更新套餐菜单项接口"
      parameters:
        - name: "restId"
          in: path
          required: true
          type: string
          description: "菜单项所属菜单餐厅ID"
        - name: "itemId"
          in: path
          required: true
          type: string
          description: "菜单项ID"
        - name: "更新套餐菜单项命令"
          in: body
          required: true
          schema:
            $ref: "#/definitions/UpdateSetMenuItemCommand"
      responses:
        204:
          description: "更新单品菜单项成功"
...
  /add-items-to-set-menu-item/:setMenuItemId:
    post:
      tags:
        - "管理员维护餐厅的菜单信息"
      summary: "为套餐添加单品菜单项接口"
      parameters:
        - name: "setMenuItemId"
          type: string
          in: path
          required: true
          description: "套餐菜单项ID"
        - name: "为套餐添加单品菜单项命令"
          in: body
          required: true
          schema:
            $ref: "#/definitions/AddItemsToSetMenuItemCommand"
      responses:
        204:
          description: "添加成功"
  /delete-single-menu-item/:menuItemId/of-restaurant/:restId:
    post:
      tags:
        - "管理员维护餐厅的菜单信息"
      summary: "管理员删除某个餐厅的某个单品接口"
      parameters:
        - name: "menuItemId"
          in: path
          required: true
          type: string
          description: "单品菜单项ID"
        - name: "restId"
          in: path
          required: true
          type: string
          description: "餐厅ID"
      responses:
        204:
          description: "删除成功"
        400:
          headers:
            X-BIZ-ERR-CODE:
              type: string
              description: "40007"
          description: "请求格式有误"
          schema:
            $ref: "#/definitions/EX40007"
  /delete-set-menu-item/:menuItemId/of-restaurant/:restId:
    post:
      tags:
        - "管理员维护餐厅的菜单信息"
      summary: "管理员删除某个餐厅的某个套餐接口"
      parameters:
        - name: "menuItemId"
          in: path
          required: true
          type: string
          description: "套餐菜单项ID"
        - name: "restId"
          in: path
          required: true
          type: string
          description: "餐厅ID"
      responses:
        204:
          description: "删除成功"
        400:
          headers:
            X-BIZ-ERR-CODE:
              type: string
              description: "40007"
          description: "请求格式有误"
          schema:
            $ref: "#/definitions/EX40007"
...
  /create-mr-group:
    post:
      tags:
        - "点餐员维护点餐团信息"
      summary: "点餐员创建点参团接口"
      parameters:
        - name: "点参团的创建信息"
          in: body
          required: true
          schema:
            $ref: "#/definitions/CreateMealReservationGroupDTO"
      responses:
        200:
          description: "创建成功"
          schema:
            $ref: "#/definitions/AggregateCreatedDTO"
...
  /delete-mr-group/:id:
    get:
      tags:
        - "点餐员维护点餐团信息"
      summary: "点餐员删除点参团接口"
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: "点参团ID"
      responses:
        204:
          description: "删除成功"
...
securityDefinitions:
  CMRSSecurity:
    type: apiKey
    in: header
    name: Authorization
definitions:
...
  CreateRestaurantCommand:
    type: object
    properties:
      name:
        type: string
        description: "餐厅名称"
        example: "四季餐厅"
      description:
        type: string
        description: "餐厅介绍"
        example: "四季餐厅的介绍"
    required:
      - name
      - description
  UpdateRestaurantCommand:
    type: object
    properties:
      name:
        type: string
        description: "餐厅名称"
        example: "四季餐厅"
      description:
        type: string
        description: "餐厅介绍"
        example: "四季餐厅的介绍"
  CreateSingleMenuItemCommand:
    type: object
    properties:
      restaurantId:
        type: string
        example: "f52bb20a-45f0-44c1-b424-e2637e2c2d3e"
      name:
        type: string
        example: "萝卜丸子汤"
      dishTypeId:
        type: string
        example: "1ab7d23f-5857-4d0b-9f19-dbde727c08bf"
      tasteId:
        type: string
        example: null
      price:
        type: number
        example: 15.5
    required:
      - restaurantId
      - name
      - dishTypeId
      - price
  CreateSetMenuItemCommand:
    type: object
    properties:
      restaurantId:
        type: string
        example: "f52bb20a-45f0-44c1-b424-e2637e2c2d3e"
      name:
        type: string
        example: "商务简餐"
      price:
        type: number
        description: "套餐的定价，一般会比总价低"
        example: 25.5
      subItems:
        type: array
        items:
          type: object
          properties:
            singleMenuItemId:
              type: string
              example: "7de43c37-714c-49e7-8a56-a35a29e2e552"
            quantity:
              type: integer
              example: 2
    required:
      - restaurantId
      - name
      - price
      - subItems
  UpdateSingleMenuItemCommand:
    type: object
    properties:
      name:
        type: string
        example: "西湖牛肉羹"
      dishTypeId:
        type: string
        example: "1ab7d23f-5857-4d0b-9f19-dbde727c08bf"
      tasteId:
        type: string
        example: "97ff7f12-f28b-416a-9600-f8927b0fb281"
      price:
        type: number
        example: 12.5
      onShelve:
        type: boolean
        example: true
  UpdateSetMenuItemCommand:
    type: object
    properties:
      name:
        type: string
        example: "情侣套餐"
      price:
        type: number
        example: 15.5
      onShelve:
        type: boolean
        example: false
      subItems:
        type: array
        items:
          type: object
          properties:
            singleMenuItemId:
              type: string
              example: "97ff7f12-f28b-416a-9600-f8927b0fb281"
            quantity:
              type: integer
              example: 1
...
  AggregateCreatedDTO:
    type: object
    properties:
      code:
        type: string
        example: "20000"
      data:
        type: object
        properties:
          id:
            type: string
            example: "b6bdf1c1-5094-45d8-ad1d-231e0f19b48b"
...
  AddItemsToSetMenuItemCommand:
    type: object
    properties:
      itemsId:
        type: array
        items:
          type: string
          example:
            - "fe925f17-7746-41fd-b420-1268e60b5910"
            - "dc80b746-d889-45c9-8c5e-4b35257c871d"
...
  CreateMealReservationGroupDTO:
    type: object
    properties:
      name:
        type: string
        minLength: 4
        maxLength: 8
        example: "我的点参团"
...
  EX40007:
    type: object
    properties:
      message:
        type: string
        example: "指定的菜单项不存在"
...